refactor(parser): split primary/ident.rs into cohesive submodules (§7-8)
Split the 3111-line ident.rs into a thin facade + 7 cohesive submodules:
- supply.rs: supply_method_call and body rewriting helpers
- circumfix.rs: declared_circumfix_op, parse_circumfix_rest, parse_raw_braced_regex_body
- term_literals.rs: declared_term_symbol, class_literal, whatever, keyword_literal
- predicates.rs: is_keyword, is_listop, is_expr_listop, balanced_paren_text, and other predicate helpers
- listop.rs: make_call_expr, parse_expr_listop_args, try_parse_no_paren_invocant_colon_call
- anon_sub.rs: parse_anon_sub_with_params, parse_anon_method_with_params, make_anon_method
- identifier_call.rs: the indivisible ~1400-line identifier_or_call + parse_require_expr

Pure relocation: no logic changes, no external-API changes.
All super::super:: paths rewritten to absolute crate::parser:: paths.
